ila: Adjusted sample padding
`self.bits_per_sample` was the closest power of 2 number of bites, and `self.bytes_per_sample` was derived off of that.

This caused some problems, notably that when the ILA samples hit 16/17 bits, they would be padded up to 32, giving us a full byte overhead.

This has been adjusted so that now `self.bytes_per_sample` is calculated off of the raw `sample_width` up to the nearest whole byte number of bits, and then `self.bits_per_sample` is derived off of that, meaning our worst-case sample padding overhead is now 6 bits. not a possible exponential number of padding bits due to the power of two (32 signal samples would result in 4-extra bytes of just 0's, 64 in 8, etc)